fixes in game_model + controller implemented (for game_model, ready for integration with the view side of things)
