+++ controller/controller.h
@@ -10,28 +10,31 @@ class Controller : public QObject {
 public:
     Controller(Model *model, QObject *parent = nullptr);
 
+    // gameloop controllers
     void startGame();
     void onCardSelected(int cardIndex);
     void onCardDeselected(int cardIndex);
     void onCardHovered(int cardIndex);
     void onCardsLowered();
-    void onRandomisedCardScreen(bool forPrimaryPlayer);
-    void onMakeGuess();
-    void onExitToMainMenu();
-    void addToLeaderboard(bool includeSecondaryPlayer);
+    std::pair<int, int> getRandomisedCards();
+    void onMakeGuess(int cardIndex);
+    void onExitToMainMenu(bool includeSecondaryPlayer, QString sndUsername="");
+    void addToLeaderboard(bool includeSecondaryPlayer, QString sndUsername="");
     bool getPlayerTurn();
     void invertPlayerTurn();
     void onTurnOverEnd();
     // TODO: include functions for the leaderboard, settings and game creation pages
 
 public slots:
+    // gameloop controller slots
     void onTurnOver(); // link to Model::turnOverEvent()
     void onGameOver(); // link to Model::endEvent()
     void onLowerFail(); // link to Model::lowerFail()
 
 signals: // to connect to View SLOTS, which are yet to be implemented
+    // gameloop controller signals
     void showRulesScreen();
-    void introduceRandomisedCards();
+    void introduceRandomisedCards(); // called in a View function, not here
     void showGameScreen();
     void promptLowerFailPopup();
     void showTurnOverScreen();
@@ -42,45 +45,5 @@ private:
     Model *model;
     bool isPrimaryPlayer = true;
 };
-
-void Controller::startGame() 
-{
-    model->initGame();
-}
-
-void Controller::onCardSelected(int cardIndex) 
-{
-    model->addCardToSelection(cardIndex);
-}
-
-void Controller::onCardDeselected(int cardIndex) 
-{
-    model->removeCardFromSelection(cardIndex);
-}
-
-void Controller::onCardHovered(int cardIndex) 
-{
-    std::pair<std::string, std::string> movieData = model->getData(cardIndex);
-    // ???????????
-}
-void Controller::onCardsLowered() 
-{
-    model->lowerCards();
-}
-
-bool Controller::getPlayerTurn() 
-{
-    return isPrimaryPlayer;
-}
-void Controller::invertPlayerTurn() 
-{
-    isPrimaryPlayer = !isPrimaryPlayer;
-}
-
-void Controller::onTurnOver() 
-{
-    invertPlayerTurn();
-    emit showTurnOverScreen();
-}
 
 #endif
